refactor: re-organize components folder structure
