+++ src/components/Header/Header.tsx
@@ -1,4 +1,4 @@
-import logo from '../assets/investment-calculator-logo.png';
+import logo from '../../assets/investment-calculator-logo.png';
 import classes from './Header.module.css';
 
 const Header = () => {
