+++ src/components/InvestmentResult/InvestmentResult.module.css
